# src/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/signin": {
      "post": {
        "description": "Validates user credentials. On success, sets HttpOnly `accessToken` (10 min) and `refresh_token` (7 days) cookies. If 2FA is enabled on the account, returns a temporary challenge token (`tempToken`) instead of setting session cookies.",
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful or 2FA challenge triggered",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/AuthMessageResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/TwoFactorChallengeResponseDto"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input format (empty email, invalid email syntax, or short password)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or inactive account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Sign in with email and password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/signup": {
      "post": {
        "description": "Creates a new user profile with unique username, lowercase email, and password. Enforces strong password rules (uppercase, lowercase, number/symbol, min 9 chars).",
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignUpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed on registration fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Username or email is already taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Register a new user account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/signout": {
      "delete": {
        "description": "Revokes the current JWT session from the database and clears `accessToken` and `refresh_token` cookies.",
        "operationId": "AuthController_signOut",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully signed out; session revoked and cookies cleared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessageResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required: missing or invalid accessToken cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Sign out and revoke session",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "description": "Exchanges a valid HttpOnly `refresh_token` cookie for a new 10-minute `accessToken` cookie.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token successfully refreshed; new accessToken cookie set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessageResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token missing, invalid, or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "refresh_token": []
          }
        ],
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/twofactor/enable": {
      "post": {
        "description": "Verifies the user account password and generates a new TOTP secret key and QR code data URI.",
        "operationId": "AuthController_twoFactorEnable",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "TOTP secret generated; returns QR code URI and base32 key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorEnableResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Password confirmation failed or is missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required: missing or invalid accessToken cookie\n\nInvalid password provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Two-factor authentication is already enabled for this account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Enable two-factor authentication (initiate setup)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/twofactor/access": {
      "post": {
        "description": "Validates the 6-digit TOTP code against the temporary token received from `/auth/signin`. On success, issues authentication cookies.",
        "operationId": "AuthController_accessTwoFactor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFactorAuth"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Two-factor sign-in successful; session cookies issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Code must be exactly 6 digits; tempToken must be a valid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired 2FA code or temporary token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Complete two-factor sign-in",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/twofactor/verify": {
      "post": {
        "description": "Submits the initial 6-digit TOTP code generated by the user authenticator app to activate 2FA protection on the account.",
        "operationId": "AuthController_verify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFactorCode"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Two-factor authentication successfully activated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Code must be exactly 6 numeric digits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required: missing or invalid accessToken cookie\n\nInvalid verification code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Verify and activate 2FA setup",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/twofactor/disable": {
      "post": {
        "description": "Verifies account password and disables 2FA protection, clearing the stored secret.",
        "operationId": "AuthController_twoFactorDisable",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Two-factor authentication successfully disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Password confirmation failed or is missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required: missing or invalid accessToken cookie\n\nInvalid password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Two-factor authentication is already disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Disable two-factor authentication",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "WeCook API",
    "description": "RESTful backend API specification for the WeCook recipe sharing social platform.\n\n### Authentication\nAuthentication is based on secure, HttpOnly JWT cookies:\n- **`accessToken`**: Short-lived (10 minutes) JWT sent automatically by browsers or HTTP clients on protected endpoints.\n- **`refresh_token`**: Long-lived (7 days) JWT restricted to the `/api/auth/refresh` path.\n\n### Media Storage & Static Assets\nUploaded recipe covers, step images, gallery photos, videos, and user avatars are served statically under `/uploads/`.\n\n### Localization & Multilingual Engine\nRecipes and catalogs support Italian (`it`), English (`en`), and French (`fr`), powered by LibreTranslate relational projection.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "https://app.we-cook.it:8443",
      "description": "Staging / Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "JWT Refresh Token stored in HttpOnly cookie (`refresh_token`) for the refresh endpoint"
      }
    },
    "schemas": {
      "AuthMessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Informative message describing the outcome of the operation",
            "example": "Authentication append with success"
          }
        },
        "required": [
          "message"
        ]
      },
      "TwoFactorChallengeResponseDto": {
        "type": "object",
        "properties": {
          "tempToken": {
            "type": "string",
            "description": "Temporary JWT token to exchange in /auth/twofactor/access",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "twofAuth": {
            "type": "boolean",
            "description": "Flag indicating two-factor verification is required",
            "example": true
          }
        },
        "required": [
          "tempToken",
          "twofAuth"
        ]
      },
      "SignInUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Registered user email address",
            "example": "mario.rossi@example.com",
            "maxLength": 255,
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "User account password (minimum 9 characters)",
            "example": "SuperSecret123!",
            "minLength": 9
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status code representing the error category",
            "example": 400
          },
          "timestamp": {
            "type": "string",
            "description": "ISO 8601 formatted timestamp of when the error occurred",
            "example": "2026-10-09T17:30:00.000Z"
          },
          "path": {
            "type": "string",
            "description": "Requested URL path that produced the error",
            "example": "/api/recipes/42"
          },
          "message": {
            "description": "Error description or list of validation error messages",
            "example": "Recipe not found",
            "oneOf": [
              {
                "type": "string",
                "example": "Recipe not found"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "Title must be at least 3 characters long"
                ]
              }
            ]
          }
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "message"
        ]
      },
      "SignUpUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Unique username consisting of letters, numbers, underscores, and hyphens (3-30 characters)",
            "example": "mario_rossi",
            "minLength": 3,
            "maxLength": 30,
            "pattern": "^[a-zA-Z0-9_-]+$"
          },
          "email": {
            "type": "string",
            "description": "Valid email address",
            "example": "mario.rossi@example.com",
            "maxLength": 255,
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "Strong password (9-72 chars, must contain uppercase, lowercase, and a number or symbol)",
            "example": "SecretPassw0rd!",
            "minLength": 9,
            "maxLength": 72,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*[\\d\\W_]).+$"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "SignUpResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique user identifier",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "Unique username of the registered user",
            "example": "mario_rossi"
          },
          "email": {
            "type": "string",
            "description": "Email address of the user",
            "example": "mario.rossi@example.com"
          },
          "role": {
            "description": "System role assigned to the user",
            "example": "user",
            "enum": [
              "admin",
              "moderator",
              "user"
            ],
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Account creation timestamp",
            "example": "2026-10-09T18:00:00.000Z"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "role",
          "created_at"
        ]
      },
      "ConfirmPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "User account password (minimum 9 characters)",
            "example": "SuperSecret123!",
            "minLength": 9
          }
        },
        "required": [
          "password"
        ]
      },
      "TwoFactorEnableResponseDto": {
        "type": "object",
        "properties": {
          "qrCode": {
            "type": "string",
            "description": "Data URI representation of the QR code containing OTP auth URI",
            "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
          },
          "key": {
            "type": "string",
            "description": "Base32 secret key for manual entry in authenticator application",
            "example": "JBSWY3DPEHPK3PXP"
          }
        },
        "required": [
          "qrCode",
          "key"
        ]
      },
      "TwoFactorAuth": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "6-digit numeric TOTP authentication code",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          },
          "tempToken": {
            "type": "string",
            "description": "Temporary JWT token received from /api/auth/signin challenge",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "code",
          "tempToken"
        ]
      },
      "TwoFactorCode": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "6-digit numeric TOTP authentication code",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "code"
        ]
      },
...
    }
  }
}